refactor: improve typing for custom styles
